requirements.txt
Django==4.2
djangorestframework==3.15.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
# psycopg2-binary==2.9.6   # optional if you use Postgres
# django-cors-headers==3.15.0
